Hotfix for rare sound bug, removed print from production
Added additional checks to the Volume loop to ensure the client can actually detect the source player and their pet, when setting the volume of a pass alarm. There seems to be a rare 'bug' whereby the client attempts to fetch the source player via their server ID in the client script, then collects their pet as well, and somewhere in that process one of the native returns `-1`, which indicates that the fetch failed, however, because these natives were nested inside both a `GetPlayerPed ()` and `GetEntityCoords ()`, a `-1` value is a valid argument for `GetPlayerPed ()`, as it just returns theclient's ped, the same as `PlayerPedId()`; thus the client is getting their own coordinates instead of that of the source player, causing them to hear Pass sounds at full volume, even if they are nowhere near the source player.

Also removed a testing print that was not intended for production from the volume loop.